change int_fast64_t to long long

+++ Algorithm/Meguru.cpp
@@ -1,6 +1,6 @@
-int_fast64_t ok = $1, ng = $2;
+long long ok = $1, ng = $2;
 while (abs(ok - ng) > 1) {
-    int_fast64_t mid = (ok + ng) >> 1;
+    long long mid = (ok + ng) >> 1;
     if (check(mid)) {
         ok = mid;
     } else {
