+++ Math/Prime/Divisor.cpp
@@ -1,6 +1,6 @@
-vector<int_fast64_t> divisor(const int_fast64_t n) {
-    vector<int_fast64_t> ret;
-    for (int_fast64_t i = 1; i * i <= n; ++i) {
+vector<long long> divisor(const long long n) {
+    vector<long long> ret;
+    for (long long i = 1; i * i <= n; ++i) {
         if (n % i == 0) {
             ret.emplace_back(i);
             if (i * i != n) ret.emplace_back(n / i);
@@ -13,6 +13,6 @@ vector<int_fast64_t> divisor(const int_fast64_t n) {
 /*
 ・約数列挙
     > O(√n)
-[使用例] vector<int_fast64_t> div = divisor(105);
+[使用例] vector<long long> div = divisor(105);
 [備考] 整列されていない.
 */
